+++ commons/ava_testsuite/unrequested_chit_spammer_test/unrequested_chit_spammer_test_.go
@@ -30,7 +30,7 @@ type StakingNetworkUnrequestedChitSpammerTest struct{
 }
 func (test StakingNetworkUnrequestedChitSpammerTest) Run(network networks.Network, context testsuite.TestContext) {
 	castedNetwork := network.(ava_networks.TestGeckoNetwork)
-	networkAcceptanceTimeout := time.Duration(networkAcceptanceTimeoutRatio * float64(test.GetTimeout().Nanoseconds()))
+	networkAcceptanceTimeout := time.Duration(networkAcceptanceTimeoutRatio * float64(test.GetExecutionTimeout().Nanoseconds()))
 
 	for i := 0; i < int(normalNodeServiceId); i++ {
 		byzClient, err := castedNetwork.GetGeckoClient(networks.ServiceID(i))
@@ -88,12 +88,16 @@ func (test StakingNetworkUnrequestedChitSpammerTest) GetNetworkLoader() (network
 	}
 	return getByzantineNetworkLoader(serviceIdConfigMap, test.UnrequestedChitSpammerImageName, test.NormalImageName)
 }
-func (test StakingNetworkUnrequestedChitSpammerTest) GetTimeout() time.Duration {
-	// TODO drop this when the availabilityChecker doesn't have a sleep
-	return 720 * time.Second
+
+func (test StakingNetworkUnrequestedChitSpammerTest) GetExecutionTimeout() time.Duration {
+	return 5 * time.Minute
 }
 
-
+func (test StakingNetworkUnrequestedChitSpammerTest) GetSetupBuffer() time.Duration {
+	// TODO drop this when the availabilityChecker doesn't have a sleep, because we spin up a *bunch* of nodes before test
+	//  execution starts
+	return 12 * time.Minute
+}
 
 // =============== Helper functions =============================